[CODE] Implemented Machine Learn for classification
