+++ apps/risk_rating/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render
-from apps.risk_rating import ml_classification
+from apps.risk_rating import ml_classifier
 
 
 def risk_rating_view(request):
@@ -25,7 +25,7 @@ def risk_rating_view(request):
         }
 
         print(patient)
-        print(ml_classification.classify_csv())
+        print(ml_classifier.classify_csv())
 
     return render(request, 'risk_rating/risk_rating.html')
